Revert "reduce propdrilling wip"
This reverts commit 9887f6d3c476d90522273fd8c46fe9f62a6c2839.
