+++ apps/portal/app/components/create-claim/create-claim-form.tsx
@@ -1,44 +1,24 @@
-import { useEffect, useReducer } from 'react'
+import { useEffect } from 'react'
 
 import {
   DialogDescription,
   DialogHeader,
   DialogTitle,
   Text,
-  toast,
   TransactionStatus,
   TransactionStatusType,
 } from '@0xintuition/1ui'
-import { ClaimPresenter, IdentityPresenter } from '@0xintuition/api'
 
-import { SubmissionResult } from '@conform-to/react'
-import { multivaultAbi } from '@lib/abis/multivault'
-import { useCreateTriple } from '@lib/hooks/useCreateTriple'
-import { useIdentitySelection } from '@lib/hooks/useIdentitySelection'
-import { useLoaderFetcher } from '@lib/hooks/useLoaderFetcher'
 import { usePopoverStates } from '@lib/hooks/usePopoverStates'
-import {
-  initialTransactionState,
-  transactionReducer,
-} from '@lib/hooks/useTransactionReducer'
-import { FetcherWithComponents, useFetcher } from '@remix-run/react'
-import { CreateClaimLoaderData } from '@routes/resources+/create-claim'
-import { TagLoaderData } from '@routes/resources+/tag'
-import { useQueryClient } from '@tanstack/react-query'
-import {
-  CREATE_CLAIM_RESOURCE_ROUTE,
-  GENERIC_ERROR_MSG,
-  MULTIVAULT_CONTRACT_ADDRESS,
-  SEARCH_IDENTITIES_RESOURCE_ROUTE,
-} from 'consts'
+import { FetcherWithComponents } from '@remix-run/react'
+import { CreateClaimFeesType } from '@routes/resources+/create-claim'
 import { CLAIM_ACTIONS } from 'consts/claims'
-import { formatUnits, parseUnits } from 'viem'
-import { useBalance, useBlockNumber, usePublicClient } from 'wagmi'
+import { TransactionActionType, TransactionStateType } from 'types'
 
-import useClaimState from '../../lib/hooks/useClaimState'
+import { ClaimAction, ClaimState } from '../../lib/hooks/useClaimState'
+import { useIdentitySelection } from '../../lib/hooks/useIdentitySelection'
 import ErrorList from '../error-list'
 import { TransactionState } from '../transaction-state'
-import CreateClaimButton from './create-claim-button'
 import CreateClaimInput from './create-claim-input'
 import CreateClaimReview from './create-claim-review'
 import { IdentityPopovers } from './identity-popover'
@@ -46,252 +26,48 @@ import { IdentityPopovers } from './identity-popover'
 interface CreateClaimFormProps {
   fetchReval: FetcherWithComponents<unknown>
   formRef: React.MutableRefObject<null>
+  state: TransactionStateType
+  dispatch: React.Dispatch<TransactionActionType>
   wallet: string
-  onClose: () => void
-}
-
-export interface OffChainClaimFetcherData {
-  success: 'success' | 'error'
-  claim: ClaimPresenter
-  submission: SubmissionResult<string[]> | null
+  walletBalance: string
+  claimFetcher: FetcherWithComponents<unknown>
+  claimState: ClaimState
+  claimDispatch: React.Dispatch<ClaimAction>
+  fees: CreateClaimFeesType
+  isLoading?: boolean
 }
 
 export function CreateClaimForm({
   fetchReval,
   formRef,
+  state,
+  dispatch,
   wallet,
-  onClose,
+  walletBalance,
+  claimFetcher,
+  claimState,
+  claimDispatch,
+  fees,
+  isLoading = false,
 }: CreateClaimFormProps) {
-  const [transactionState, transactionDispatch] = useReducer(
-    transactionReducer,
-    initialTransactionState,
-  )
-  const { claimState, claimDispatch } = useClaimState()
+  const popoverStates = usePopoverStates()
+
   const {
-    setSearchQuery,
-    setIdentities,
     selectedIdentities,
+    setSearchQuery,
     identities,
     handleIdentitySelection,
   } = useIdentitySelection()
-  const popoverStates = usePopoverStates()
-  const publicClient = usePublicClient()
-  const claimFetcher = useFetcher<OffChainClaimFetcherData | null>()
-  const feeFetcher = useLoaderFetcher<CreateClaimLoaderData>(
-    CREATE_CLAIM_RESOURCE_ROUTE,
-  )
-  const identitiesFetcher = useFetcher<IdentityPresenter[] | null>()
-  const claimChecker = useFetcher<TagLoaderData | null>()
-
-  const queryClient = useQueryClient()
-  const { data: blockNumber } = useBlockNumber({ watch: true })
-  const { data: balance, queryKey } = useBalance({
-    address: wallet as `0x${string}`,
-  })
-
-  const walletBalance = formatUnits(balance?.value ?? 0n, 18)
-
-  const {
-    writeContractAsync: writeCreateTriple,
-    awaitingWalletConfirmation,
-    awaitingOnChainConfirmation,
-    isError,
-    reset,
-  } = useCreateTriple()
-
-  const { fees } = (feeFetcher.data as CreateClaimLoaderData) ?? {}
-
-  const isLoading =
-    !!awaitingWalletConfirmation ||
-    !!awaitingOnChainConfirmation ||
-    transactionState.status === 'confirm' ||
-    transactionState.status === 'transaction-pending' ||
-    transactionState.status === 'transaction-confirmed'
 
   useEffect(() => {
-    if (claimFetcher.data?.claim && !claimState.createdClaim) {
-      claimDispatch({
-        type: CLAIM_ACTIONS.SET_CREATED_CLAIM,
-        payload: claimFetcher.data.claim,
-      })
-    }
-  }, [claimFetcher.data, claimFetcher.state, claimState.createdClaim])
-
-  useEffect(() => {
-    if (claimState.createdClaim) {
-      handleCreateTriple()
-    }
-  }, [claimState.createdClaim])
-
-  useEffect(() => {
-    if (awaitingWalletConfirmation) {
-      transactionDispatch({ type: 'APPROVE_TRANSACTION' })
-    }
-    if (awaitingOnChainConfirmation) {
-      transactionDispatch({ type: 'TRANSACTION_PENDING' })
-    }
-    if (isError) {
-      transactionDispatch({
-        type: 'TRANSACTION_ERROR',
-        error: 'Error processing transaction',
-      })
-    }
-  }, [awaitingWalletConfirmation, awaitingOnChainConfirmation, isError])
-
-  useEffect(() => {
-    if (isError) {
-      reset()
-    }
-  }, [isError, reset])
-
-  useEffect(() => {
-    if (blockNumber && blockNumber % 5n === 0n) {
-      queryClient.invalidateQueries({ queryKey })
-    }
-  }, [blockNumber, queryClient, queryKey])
-
-  useEffect(() => {
-    if (claimState.searchQuery) {
-      const searchParam = `?search=${encodeURIComponent(claimState.searchQuery)}`
-      identitiesFetcher.load(
-        `${SEARCH_IDENTITIES_RESOURCE_ROUTE}${searchParam}`,
-      )
-    }
-  }, [claimState.searchQuery])
-
-  useEffect(() => {
-    if (identitiesFetcher.data) {
-      claimDispatch({
-        type: CLAIM_ACTIONS.SET_IDENTITIES,
-        payload: identitiesFetcher.data,
-      })
-    }
-  }, [identitiesFetcher.data])
-
-  useEffect(() => {
-    if (
-      claimState.selectedIdentities.subject &&
-      claimState.selectedIdentities.predicate &&
-      claimState.selectedIdentities.object &&
-      !claimState.claimExists
-    ) {
-      claimChecker.load(
-        `/resources/tag?subjectId=${claimState.selectedIdentities.subject.vault_id}&predicateId=${claimState.selectedIdentities.predicate.vault_id}&objectId=${claimState.selectedIdentities.object.vault_id}`,
-      )
-    }
-  }, [claimState.selectedIdentities])
-
-  useEffect(() => {
-    if (claimChecker.data) {
-      claimDispatch({
-        type: CLAIM_ACTIONS.SET_CLAIM_EXISTS,
-        payload: claimChecker.data.result !== '0',
-      })
-    }
-  }, [claimChecker.data])
-
-  const handleCreateTriple = async () => {
-    try {
-      const txHash = await writeCreateTriple({
-        address: MULTIVAULT_CONTRACT_ADDRESS,
-        abi: multivaultAbi,
-        functionName: 'createTriple',
-        args: [
-          claimState.selectedIdentities.subject?.vault_id,
-          claimState.selectedIdentities.predicate?.vault_id,
-          claimState.selectedIdentities.object?.vault_id,
-        ],
-        value:
-          (fees?.tripleCost ? BigInt(fees.tripleCost) : 0n) +
-          parseUnits(
-            claimState.initialDeposit && claimState.initialDeposit !== ''
-              ? claimState.initialDeposit
-              : '0',
-            18,
-          ),
-      })
-
-      if (publicClient && txHash) {
-        transactionDispatch({ type: 'TRANSACTION_PENDING' })
-        const receipt = await publicClient.waitForTransactionReceipt({
-          hash: txHash,
-        })
-
-        if (receipt.status === 'reverted') {
-          transactionDispatch({
-            type: 'TRANSACTION_ERROR',
-            error: 'Transaction reverted',
-          })
-          toast.error('Transaction reverted. Please try again.')
-        } else {
-          transactionDispatch({
-            type: 'TRANSACTION_COMPLETE',
-            txHash,
-            txReceipt: receipt,
-          })
-          fetchReval.submit(formRef.current, {
-            method: 'POST',
-          })
-        }
-      }
-    } catch (error) {
-      console.error('error', error)
-      if (error instanceof Error) {
-        let errorMessage = 'Error in onchain transaction.'
-        if (error.message.includes('insufficient')) {
-          errorMessage =
-            'Insufficient funds. Please add more ETH to your wallet and try again.'
-        }
-        if (error.message.includes('rejected')) {
-          errorMessage = 'Transaction rejected. Try again when you are ready.'
-        }
-        transactionDispatch({
-          type: 'TRANSACTION_ERROR',
-          error: errorMessage,
-        })
-        toast.error(GENERIC_ERROR_MSG)
-      }
-    }
-  }
-  const handleCreateButtonClick = async () => {
-    transactionDispatch({ type: 'APPROVE_TRANSACTION' })
-    claimDispatch({ type: CLAIM_ACTIONS.SET_CREATED_CLAIM, payload: null })
-    const formData = new FormData()
-    formData.append(
-      'subject_id',
-      claimState.selectedIdentities.subject?.id ?? '',
-    )
-    formData.append(
-      'predicate_id',
-      claimState.selectedIdentities.predicate?.id ?? '',
-    )
-    formData.append('object_id', claimState.selectedIdentities.object?.id ?? '')
-    formData.append('initial_deposit', claimState.initialDeposit)
-    if (transactionState.status === 'review-transaction') {
-      claimFetcher.submit(formData, {
-        action: '/actions/create-claim',
-        method: 'post',
-      })
-    }
-  }
-
-  const handleClose = () => {
-    onClose()
-    setTimeout(() => {
-      transactionDispatch({ type: 'START_TRANSACTION' })
-      reset()
-      claimDispatch({ type: CLAIM_ACTIONS.RESET })
-      claimFetcher.data = null
-      claimChecker.data = null
-      identitiesFetcher.data = null
-      setSearchQuery('')
-      setIdentities([])
-      claimDispatch({ type: CLAIM_ACTIONS.SET_CREATED_CLAIM, payload: null })
-    }, 500)
-  }
+    claimDispatch({
+      type: CLAIM_ACTIONS.SET_SELECTED_IDENTITIES,
+      payload: selectedIdentities,
+    })
+  }, [selectedIdentities, claimDispatch])
 
   const renderContent = () => {
-    switch (transactionState.status) {
+    switch (state.status) {
       case TransactionStatus.idle:
         return (
           <>
@@ -316,6 +92,7 @@ export function CreateClaimForm({
                     identities={identities}
                     handleIdentitySelection={handleIdentitySelection}
                     setSearchQuery={setSearchQuery}
+                    popoverStates={popoverStates}
                     handleInput={async (event) => {
                       const query = event.currentTarget.value
                       if (query.trim() !== '') {
@@ -325,7 +102,6 @@ export function CreateClaimForm({
                         })
                       }
                     }}
-                    popoverStates={popoverStates}
                   />
                   <CreateClaimInput
                     val={claimState.initialDeposit}
@@ -360,7 +136,7 @@ export function CreateClaimForm({
         return (
           <div className="h-full flex flex-col">
             <CreateClaimReview
-              dispatch={transactionDispatch}
+              dispatch={dispatch}
               selectedIdentities={selectedIdentities}
               initialDeposit={claimState.initialDeposit}
               fees={fees}
@@ -371,8 +147,8 @@ export function CreateClaimForm({
         return (
           <div className="flex flex-col items-center justify-center flex-grow">
             <TransactionState
-              status={transactionState.status as TransactionStatusType}
-              txHash={transactionState.txHash}
+              status={state.status as TransactionStatusType}
+              txHash={state.txHash}
               type="claim"
             />
           </div>
@@ -391,19 +167,6 @@ export function CreateClaimForm({
         <input type="hidden" name="eventName" value="create" />
       </fetchReval.Form>
       {renderContent()}
-      {transactionState.status !== 'transaction-pending' &&
-        transactionState.status !== 'awaiting' &&
-        transactionState.status !== 'confirm' && (
-          <CreateClaimButton
-            handleAction={handleCreateButtonClick}
-            handleClose={handleClose}
-            dispatch={transactionDispatch}
-            state={transactionState}
-            claim_id={claimState.createdClaim?.claim_id ?? ''}
-            selectedIdentities={claimState.selectedIdentities}
-            claimExists={claimState.claimExists}
-          />
-        )}
     </div>
   )
 }
